constants_refactor split constants.py into constants directory, split into categories (keys, enums, sprites, images, stats, weapons, colors, and misc stayed in constants.py)
