+++ src/action/ship_config/change_select.py
@@ -4,7 +4,7 @@ from typing import TYPE_CHECKING
 
 from action.base import Action
 from custom_exceptions import Impossible
-from enums import GameStates
+from constants.enums import GameStates
 
 if TYPE_CHECKING:
     from entity import Actor
