+++ src/action/ship_config/configure.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING
 from action.base import Action
 from action.ship_config.exit import ExitMenuAction
 from custom_exceptions import Impossible
-from enums import GameStates
+from constants.enums import GameStates
 
 if TYPE_CHECKING:
     from entity import Actor
