+++ src/components/fighter.py
@@ -1,7 +1,8 @@
 from components.base import BaseComponent
-from constants import colors, move_elevations
+from constants.constants import move_elevations
+from constants.colors import colors
 from entity import Actor
-from enums import GameStates, RenderOrder
+from constants.enums import GameStates, RenderOrder
 from event_handlers.player_dead import GameOverEventHandler
