+++ src/constants/colors.py
@@ -1,0 +1,34 @@
+colors = {
+    'white': (255, 255, 255),
+    'lt_gray': (200, 200, 200),
+    'dk_gray': (100, 100, 100),
+    'dark': (25, 25, 25),
+    'dark_green': (25, 50, 25),
+    'black': (0, 0, 0),
+    'pink': (225, 175, 175),
+    'red': (250, 50, 50),
+    'orange': (225, 150, 100),
+    'bar_filled': (0, 100, 0),
+    'bar_empty': (75, 25, 25),
+    'invalid': (255, 255, 0),
+    'gray': (125, 125, 125),
+    'error': (255, 75, 75),
+    'green': (0, 255, 0),
+    'ocean': (0, 0, 175),
+    'water': (0, 50, 200),
+    'shallows': (0, 75, 225),
+    'beach': (225, 200, 125),
+    'grass': (50, 175, 50),
+    'jungle': (0, 125, 0),
+    'mountain': (225, 225, 225),
+    'volcano': (200, 0, 0),
+    'aqua': (0, 125, 255),
+    'cyan': (0, 255, 255),
+    'violet': (125, 0, 255),
+    'purple': (255, 0, 255),
+    'coral': (255, 115, 200),
+    'rocks': (225, 225, 225),
+    'seaweed': (50, 175, 50),
+    'sandbar': (225, 200, 125),
+    'minefield': (200, 0, 0),
+}
