+++ src/constants/weapons.py
@@ -1,0 +1,58 @@
+weapons = {
+    'ballista': {'hp': 3,
+                 'defense': 1,
+                 'range': 3,
+                 'power': 3,
+                 'cooldown': 4,
+                 'ammo': 'bolts',
+                 },
+    'heavy ballista': {'hp': 4,
+                       'defense': 2,
+                       'range': 3,
+                       'power': 4,
+                       'cooldown': 4,
+                       'ammo': 'bolts',
+                       },
+    'repeating ballista': {'hp': 3,
+                           'defense': 1,
+                           'range': 3,
+                           'power': 3,
+                           'cooldown': 3,
+                           'ammo': 'bolts',
+                           },
+    'sniper ballista': {'hp': 3,
+                        'defense': 1,
+                        'range': 4,
+                        'power': 3,
+                        'cooldown': 4,
+                        'ammo': 'bolts',
+                        },
+    'cannon': {'hp': 5,
+               'defense': 2,
+               'range': 4,
+               'power': 4,
+               'cooldown': 5,
+               'ammo': 'cannonballs',
+               },
+    'organ gun': {'hp': 4,
+                  'defense': 2,
+                  'range': 4,
+                  'power': 4,
+                  'cooldown': 4,
+                  'ammo': 'cannonballs',
+                  },
+    'heavy cannon': {'hp': 6,
+                     'defense': 3,
+                     'range': 4,
+                     'power': 6,
+                     'cooldown': 6,
+                     'ammo': 'cannonballs',
+                     },
+    'long guns': {'hp': 5,
+                  'defense': 2,
+                  'range': 5,
+                  'power': 4,
+                  'cooldown': 5,
+                  'ammo': 'cannonballs',
+                  },
+}
