+++ src/entity_factory.py
@@ -5,7 +5,8 @@ from components.crew import Crew
 from components.fighter import Fighter
 from components.sails import Sails
 from components.view import View
-from constants import animation_speed, flicker_timer, sprite_count, move_elevations
+from constants.constants import move_elevations
+from constants.sprites import animation_speed, flicker_timer, sprite_count
 from entity import Actor, Entity
 from sprite import Sprite
