+++ src/game_map.py
@@ -4,8 +4,8 @@ from queue import Queue
 from random import randint
 from typing import Iterable, List, Tuple, Optional, Set, Dict, TYPE_CHECKING
 
-from constants import colors
-from enums import Conditions, Elevation
+from constants.colors import colors
+from constants.enums import Conditions, Elevation
 from tile import Terrain
 from utilities import Hex, cube_directions, cube_add, cube_to_hex, \
     hex_to_cube, cube_neighbor, cube_line_draw, get_distance
